Create Code for the game
This is the code for wheel of fortune. The user has to guess a word in 10 attempts or he/she loses. Currently, there are only 25 words which can potentially be what the user has to guess. I would like to generate a random word from all of the English dictionary for the user to guess. Can someone please help me?